Start considering multi-threaded use
It might be that multiple threads are using a SirenEntity. In that
scenario, multiple threads must not share an Action in order to
avoid overriding each others data.
This is a first shot at tackling this issue, will probably need to
be re-evaluated quite heavily. Re-designing the API might be more
appropriate.